requirements.txt
# AI Agents
uagents==0.22.10
uagents-core==0.3.11

# LLM API
anthropic==0.39.0

# Databases
chromadb==0.5.23
sentence-transformers==3.3.1

# API Server
flask==3.0.0
flask-cors==4.0.0

# HTTP Requests
requests==2.32.5

# Environment
python-dotenv==1.0.1

# Utilities
pydantic==2.12.3
